refactor: aggregatorに単一エントリポイントrunAggregationを導入
bridgeからaggregatorへのimportを6個から1関数+1型に簡素化。
集計オーケストレーションをaggregator側に集約。
