+++ src/lib/duckdbBridge.ts
@@ -4,14 +4,7 @@ import mvp_worker from "@duckdb/duckdb-wasm/dist/duckdb-browser-mvp.worker.js?ur
 import duckdb_wasm_eh from "@duckdb/duckdb-wasm/dist/duckdb-eh.wasm?url";
 import eh_worker from "@duckdb/duckdb-wasm/dist/duckdb-browser-eh.worker.js?url";
 import type { AggResult } from "./aggregate";
-import {
-  computeTotalN,
-  fetchCrossHeaders,
-  aggregateSA,
-  aggregateMA,
-  groupMAColumns,
-  type CrossHeaderCache,
-} from "./duckdbAggregator";
+import { runAggregation, type AggPayload } from "./duckdbAggregator";
 
 type DuckStatus = "loading" | "ready" | "error";
 
@@ -72,11 +65,7 @@ function serializeToCSV(data: Record<string, string>[], cols: string[]): string 
 
 export async function runDuckDBAggregation(payload: {
   data: Record<string, string>[];
-  columns: Array<{ name: string; type: "sa" | "ma"; ma_group?: string }>;
-  weight_col: string;
-  mode: "gt" | "cross";
-  cross_cols: string[];
-}): Promise<AggResult[]> {
+} & AggPayload): Promise<AggResult[]> {
   if (!db || status !== "ready") throw new Error("DuckDB is not ready");
 
   const allColNames = [
@@ -97,31 +86,7 @@ export async function runDuckDBAggregation(payload: {
        SELECT * FROM read_csv('survey.csv', all_varchar=true)`
     );
 
-    const totalN = await computeTotalN(conn, payload.weight_col);
-    const crossCols = payload.mode === "cross" ? payload.cross_cols : [];
-    const results: AggResult[] = [];
-
-    // クロス軸ヘッダーを事前に1回だけ取得してキャッシュ
-    const crossHeaderCache =
-      crossCols.length > 0
-        ? await fetchCrossHeaders(conn, crossCols, payload.weight_col)
-        : (new Map() as CrossHeaderCache);
-
-    const saCols = payload.columns.filter((c) => c.type === "sa");
-    for (const col of saCols) {
-      results.push(
-        await aggregateSA(conn, col.name, totalN, payload.weight_col, crossCols, crossHeaderCache)
-      );
-    }
-
-    const maGroups = groupMAColumns(payload.columns);
-    for (const [prefix, cols] of Object.entries(maGroups)) {
-      results.push(
-        await aggregateMA(conn, prefix, cols, totalN, payload.weight_col)
-      );
-    }
-
-    return results;
+    return await runAggregation(conn, payload);
   } finally {
     await conn.close();
     await db.dropFile("survey.csv");
